Refactored the camera code. Also libdc1394 v1/v2 now supports selection of camera via guid (pass as the channel argument in the node).
All cameras now adhere to the idea that the construction of a camera object only succeeds if the camera could be opened succesfully. This way m_bCameraIsAvailable could me removed. The constructor throws an exception in case of problems.

User code has been migrated to use the factory getCamera in Camera.cpp. The factory will return a FakeCamera if the selected constructor fails and log the exception.


svn path=/trunk/libavg/; revision=3527